delete test controller and routing mypage_path
