+++ app/assets/javascripts/mypage.coffee
