+++ app/controllers/mypage_controller.rb
@@ -1,0 +1,25 @@
+class MypageController < ApplicationController
+    def mypage
+
+    end
+
+    def put_up_goods_on_sale
+
+    end
+
+    def purchased_on_deal
+
+    end
+
+    def edit_profile
+
+    end
+
+    def credit_enter
+
+    end
+
+    def make_sure_user_info
+        
+    end
+end
