--- app/controllers/tests_controller.rb
@@ -1,66 +1,0 @@
-class TestsController < ApplicationController
-    def index
-    end
-
-    def signup
-    end
-
-    def signup1
-    end
-
-    def signup2
-    end
-
-    def signup3
-    end
-
-    def signup4
-    end
-
-    def signup5
-    end
-
-    def user_login
-    end
-
-    def exhibit
-    end
-
-    def logout
-    end
-
-    def product_details
-    end
-
-    def credit_enter
-    end
-
-    def credit_register
-
-    end
-
-    def make_sure_user_info
-        
-    end
-
-    def edit_profile
-        
-    end
-
-    def purchase_confirmation
-        
-    end
-
-    def put_up_goods_on_sale
-
-    end
-
-    def purchased_on_deal
-
-    end
-
-    def mypage
-
-    end
-    
-end
