+++ app/helpers/mypage_helper.rb
@@ -1,0 +1,2 @@
+module MypageHelper
+end
