+++ app/views/shared/_profile-sidebar.html.haml
@@ -2,7 +2,7 @@
   %nav.mypage-nav
     %ul.mypage-nav__list
       %li
-        %a{href: tests_mypage_path, class: 'mypage-nav__list__item active'} 
+        %a{href: mypage_path, class: 'mypage-nav__list__item active'} 
           マイページ
           %i.fas.fa-chevron-right
       %li
@@ -22,7 +22,7 @@
           出品する
           %i.fas.fa-chevron-right
       %li
-        %a{href: tests_put_up_goods_on_sale_path, class: 'mypage-nav__list__item'} 
+        %a{href: mypage_put_up_goods_on_sale_path, class: 'mypage-nav__list__item'} 
           出品した商品-出品中
           %i.fas.fa-chevron-right
       %li
@@ -34,7 +34,7 @@
           出品した商品-売却済み
           %i.fas.fa-chevron-right
       %li
-        %a{href: tests_purchased_on_deal_path , class: 'mypage-nav__list__item'}
+        %a{href: mypage_purchased_on_deal_path , class: 'mypage-nav__list__item'}
           出品した商品-取引中
           %i.fas.fa-chevron-right
       %li
@@ -72,7 +72,7 @@
     %h3.mypage-nav__head 設定
     %ul.mypage-nav__list
       %li
-        %a{href: tests_edit_profile_path, class: 'mypage-nav__list__item'} 
+        %a{href: mypage_edit_profile_path, class: 'mypage-nav__list__item'} 
           プロフィール
           %i.fas.fa-chevron-right
       %li
@@ -80,7 +80,7 @@
           発送元・お届け先住所変更
           %i.fas.fa-chevron-right
       %li
-        %a{href: tests_credit_enter_path , class: 'mypage-nav__list__item'} 
+        %a{href: mypage_credit_enter_path , class: 'mypage-nav__list__item'} 
           支払い方法
           %i.fas.fa-chevron-right
       %li
@@ -88,7 +88,7 @@
           メール/パスワード
           %i.fas.fa-chevron-right
       %li
-        %a{href: tests_make_sure_user_info_path, class: 'mypage-nav__list__item'} 
+        %a{href: mypage_make_sure_user_info_path, class: 'mypage-nav__list__item'} 
           本人情報
           %i.fas.fa-chevron-right
       %li
