+++ app/views/shared/_top_page_middle.html.haml
@@ -22,7 +22,7 @@
           .middle__main__block__item__top
             %h3.middle__main__block__item__top__left
               レディース新着アイテム
-              =link_to tests_product_details_path, class:'middle__main__block__item__top__right' do
+              =link_to '#', class:'middle__main__block__item__top__right' do
                 もっと見る
               %i.fas.fa-chevron-right
