--- app/views/tests/credit_enter.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/credit-enter'
