--- app/views/tests/credit_register.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/credit-register'
