--- app/views/tests/edit_profile.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/edit-profile'
