--- app/views/tests/exhibit.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/exhibit'
