--- app/views/tests/index.html.haml
@@ -1,3 +1,0 @@
-= render 'shared/purchase_completed'
-
-
