--- app/views/tests/logout.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/logout'
