--- app/views/tests/make_sure_user_info.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/make_sure_user_info'
