--- app/views/tests/product_details.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/product-details'
