--- app/views/tests/purchase_confirmation.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/purchase-confirmation'
