--- app/views/tests/purchased_on_deal.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/purchased_on_deal'
