--- app/views/tests/put_up_goods_on_sale.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/put_up_goods_on_sale'
