--- app/views/tests/signup.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/signup'
