--- app/views/tests/signup2.html.haml
@@ -1,1 +1,0 @@
-=render 'shared/signup2'
