--- app/views/tests/signup3.html.haml
@@ -1,1 +1,0 @@
-=render 'shared/signup3'
