--- app/views/tests/signup4.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/signup4'
