--- app/views/tests/signup5.html.haml
@@ -1,1 +1,0 @@
-=render 'shared/signup5'
