--- app/views/tests/user_login.html.haml
@@ -1,1 +1,0 @@
-= render 'shared/user_login'
