+++ spec/controllers/mypage_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe MypageController, type: :controller do
+
+end
