--- spec/controllers/tests_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe TestsController, type: :controller do
-
-end
