+++ spec/helpers/mypage_helper_spec.rb
@@ -1,15 +1,15 @@
 require 'rails_helper'
 
 # Specs in this file have access to a helper object that includes
-# the TestsHelper. For example:
+# the MypageHelper. For example:
 #
-# describe TestsHelper do
+# describe MypageHelper do
 #   describe "string concat" do
 #     it "concats two strings with spaces" do
 #       expect(helper.concat_strings("this","that")).to eq("this that")
 #     end
 #   end
 # end
-RSpec.describe TestsHelper, type: :helper do
+RSpec.describe MypageHelper, type: :helper do
   pending "add some examples to (or delete) #{__FILE__}"
 end
